Rename CLI mods for consistency and discoverability
Also add the dependency on Anker, because we are going to add the
options to set up the Anker program.
