+++ cli/Cargo.toml
@@ -8,6 +8,7 @@ version = "1.1.0"
 
 [dependencies]
 anchor-lang = "0.13.2"
+anker = { path = "../anker", features = ["no-entrypoint"] }
 bincode = "1.3.1"
 borsh = "0.9.1"
 bs58 = "0.4.0"
